+++ src/homematch/__init__.py
@@ -1,5 +1,5 @@
-from ._model import LLM as LLM
-from ._model import Embeddings as Embeddings
+from ._models import LLM as LLM
+from ._models import Embeddings as Embeddings
 
 TEXT_FIELDS = ["description", "neighborhood_description"]
 METADATA_FIELDS = ["neighborhood", "price", "bedrooms", "bathrooms", "size"]
